Update issue 2
Added a 'make clean' option to the makefile 

Added some options in the FFLAGS of neoart.mk. Not all are 
needed but the one that uses real8 for the standard real 
seems necessary since otherwise the code can mix real4 
with real8. This is more or less a bug that needs to be 
removed. Will open a new issue