Update issue 1,2,3
Made some changes to the FFLAGS in neoart.mk. The previous flags 
generated a not proper functioning code. Only flag remaining is 
the real --> real8 flag 

Changed the Make file to include a test as option. It generates 
a program test in the tst directory. The output from the test 
procedure is stored in test_output, and compared with 
test_output_reference. If the diff produces a result, something 
is wrong. 

Updated TEST1.f which is currently the only test routine called
from the program test